Function winnow::ascii::alpha0

pub fn alpha0<I, E: ParserError<I>>(
    input: &mut I
) -> PResult<<I as Stream>::Slice, E>where
    I: StreamIsPartial + Stream,
    <I as Stream>::Token: AsChar,

Recognizes zero or more lowercase and uppercase ASCII alphabetic characters: 'a'..='z', 'A'..='Z'

Complete version: Will return the whole input if no terminating token is found (a non alphabetic character).

Partial version: Will return Err(winnow::error::ErrMode::Incomplete(_)) if there’s not enough input data, or if no terminating token is found (a non alphabetic character).

Example

fn parser<'s>(input: &mut &'s str) -> PResult<&'s str, InputError<&'s str>> {
    alpha0.parse_next(input)
}

assert_eq!(parser.parse_peek("ab1c"), Ok(("1c", "ab")));
assert_eq!(parser.parse_peek("1c"), Ok(("1c", "")));
assert_eq!(parser.parse_peek(""), Ok(("", "")));

assert_eq!(alpha0::<_, InputError<_>>.parse_peek(Partial::new("ab1c")), Ok((Partial::new("1c"), "ab")));
assert_eq!(alpha0::<_, InputError<_>>.parse_peek(Partial::new("1c")), Ok((Partial::new("1c"), "")));
assert_eq!(alpha0::<_, InputError<_>>.parse_peek(Partial::new("")), Err(ErrMode::Incomplete(Needed::new(1))));