Function winnow::ascii::oct_digit0

pub fn oct_digit0<I, E: ParseError<I>>(
    input: I
) -> IResult<I, <I as Stream>::Slice, E>where
    I: StreamIsPartial,
    I: Stream,
    <I as Stream>::Token: AsChar,

Recognizes zero or more octal characters: 0-7

Complete version: Will return the whole input if no terminating token is found (a non octal digit character).

Partial version: Will return Err(winnow::error::ErrMode::Incomplete(_)) if there’s not enough input data, or if no terminating token is found (a non octal digit character).

Example

fn parser(input: &str) -> IResult<&str, &str> {
    oct_digit0(input)
}

assert_eq!(parser("21cZ"), Ok(("cZ", "21")));
assert_eq!(parser("Z21c"), Ok(("Z21c", "")));
assert_eq!(parser(""), Ok(("", "")));

assert_eq!(oct_digit0::<_, Error<_>>(Partial::new("21cZ")), Ok((Partial::new("cZ"), "21")));
assert_eq!(oct_digit0::<_, Error<_>>(Partial::new("Z21c")), Ok((Partial::new("Z21c"), "")));
assert_eq!(oct_digit0::<_, Error<_>>(Partial::new("")), Err(ErrMode::Incomplete(Needed::new(1))));