Function winnow::ascii::space1

pub fn space1<Input, Error>(
    input: &mut Input
) -> PResult<<Input as Stream>::Slice, Error>where
    Input: StreamIsPartial + Stream,
    <Input as Stream>::Token: AsChar,
    Error: ParserError<Input>,

Recognizes zero or more spaces and tabs.

Complete version: Will return the whole input if no terminating token is found (a non space character).

[Partial version][crate::_topic::partial]: Will return Err(winnow::error::ErrMode::Incomplete(_)) if there’s not enough input data, or if no terminating token is found (a non space character).

Effective Signature

Assuming you are parsing a &str Stream:

pub fn space1<'i>(input: &mut &'i str) -> PResult<&'i str>

Example

fn parser<'s>(input: &mut &'s str) -> PResult<&'s str, InputError<&'s str>> {
    space1.parse_next(input)
}

assert_eq!(parser.parse_peek(" \t21c"), Ok(("21c", " \t")));
assert_eq!(parser.parse_peek("H2"), Err(ErrMode::Backtrack(InputError::new("H2", ErrorKind::Slice))));
assert_eq!(parser.parse_peek(""), Err(ErrMode::Backtrack(InputError::new("", ErrorKind::Slice))));

assert_eq!(space1::<_, InputError<_>>.parse_peek(Partial::new(" \t21c")), Ok((Partial::new("21c"), " \t")));
assert_eq!(space1::<_, InputError<_>>.parse_peek(Partial::new("H2")), Err(ErrMode::Backtrack(InputError::new(Partial::new("H2"), ErrorKind::Slice))));
assert_eq!(space1::<_, InputError<_>>.parse_peek(Partial::new("")), Err(ErrMode::Incomplete(Needed::new(1))));