Function winnow::combinator::repeat

pub fn repeat<I, O, C, E, F>(
    range: impl Into<Range>,
    f: F
) -> impl Parser<I, C, E>where
    I: Stream,
    C: Accumulate<O>,
    F: Parser<I, O, E>,
    E: ParseError<I>,

Accumulate the output of a parser into a container, like Vec

This stops before n when the parser returns ErrMode::Backtrack. To instead chain an error up, see cut_err.

Arguments

m The minimum number of iterations.

n The maximum number of iterations.

f The parser to apply.

To recognize a series of tokens, Accumulate into a () and then Parser::recognize.

Warning: If the parser passed to repeat accepts empty inputs (like alpha0 or digit0), repeat will return an error, to prevent going into an infinite loop.

Example

use winnow::combinator::repeat;
use winnow::token::tag;

fn parser(s: &str) -> IResult<&str, Vec<&str>> {
  repeat(0..=2, "abc").parse_next(s)
}

assert_eq!(parser("abcabc"), Ok(("", vec!["abc", "abc"])));
assert_eq!(parser("abc123"), Ok(("123", vec!["abc"])));
assert_eq!(parser("123123"), Ok(("123123", vec![])));
assert_eq!(parser(""), Ok(("", vec![])));
assert_eq!(parser("abcabcabc"), Ok(("abc", vec!["abc", "abc"])));