winnow::combinator

Module multi

Combinators applying their child parser multiple times

Structs

ParserIterator

Main structure associated to iterator.

Repeat

Customizable Parser implementation for repeat

Enums

State 🔒

Functions

fill

Repeats the embedded parser, filling the given slice with results.

fold_repeat0_ 🔒

fold_repeat1_ 🔒

fold_repeat_m_n_ 🔒

fold_repeat_n_ 🔒

iterator

Repeats the embedded parser, lazily returning the results

repeat

Accumulate the output of a parser into a container, like Vec (bound by occurrences)

repeat_till

Accumulate the output of parser f into a container, like Vec, until the parser g produces a result (bound by occurrences).

repeat_till0_ 🔒

repeat_till_m_n_ 🔒

separated

Accumulate the output of a parser, interleaved with sep (bound by occurrences)

separated0_ 🔒

separated1_ 🔒

separated_foldl1

Alternates between two parsers, merging the results (left associative)

separated_foldr1

Alternates between two parsers, merging the results (right associative)

separated_m_n_ 🔒

separated_n_ 🔒

try_fold_m_n 🔒

verify_fold_m_n 🔒