Trait winnow::combinator::Alt

pub trait Alt<I, O, E> {
    // Required method
    fn choice(&mut self, input: &mut I) -> PResult<O, E>;
}

Helper trait for the alt() combinator.

This trait is implemented for tuples of up to 21 elements

Required Methods

fn choice(&mut self, input: &mut I) -> PResult<O, E>

Tests each parser in the tuple and returns the result of the first one that succeeds