Trait winnow::combinator::Permutation

pub trait Permutation<I, O, E> {
    // Required method
    fn permutation(&mut self, input: &mut I) -> PResult<O, E>;
}

Helper trait for the permutation() combinator.

This trait is implemented for tuples of up to 21 elements

Required Methods

fn permutation(&mut self, input: &mut I) -> PResult<O, E>

Tries to apply all parsers in the tuple in various orders until all of them succeed