Enum winnow::error::ErrMode

pub enum ErrMode<E> {
    Incomplete(Needed),
    Backtrack(E),
    Cut(E),
}

The Err enum indicates the parser was not successful

Variants

Incomplete(Needed)

There was not enough data to determine the appropriate action

More data needs to be buffered before retrying the parse.

This must only be set when the Stream is partial, like with Partial

Convert this into an Backtrack with Parser::complete_err

Backtrack(E)

The parser failed with a recoverable error (the default).

For example, a parser for json values might include a dec_uint as one case in an alt combiantor. If it fails, the next case should be tried.

Cut(E)

The parser had an unrecoverable error.

The parser was on the right branch, so directly report it to the user rather than trying other branches. You can use cut_err() combinator to switch from ErrMode::Backtrack to ErrMode::Cut.

For example, one case in an alt combinator found a unique prefix and you want any further errors parsing the case to be reported to the user.

Implementations

impl<E> ErrMode<E>

pub fn is_incomplete(&self) -> bool

Tests if the result is Incomplete

pub fn cut(self) -> Self

Prevent backtracking, bubbling the error up to the top

pub fn backtrack(self) -> Self

Enable backtracking support

pub fn map<E2, F>(self, f: F) -> ErrMode<E2>where F: FnOnce(E) -> E2,

Applies the given function to the inner error

pub fn convert<F>(self) -> ErrMode<F>where E: ErrorConvert<F>,

Automatically converts between errors if the underlying type supports it

impl<T> ErrMode<Error<T>>

pub fn map_input<U, F>(self, f: F) -> ErrMode<Error<U>>where F: FnOnce(T) -> U,

Maps ErrMode<Error<T>> to ErrMode<Error<U>> with the given F: T -> U