Module winnow::multi

👎Deprecated since 0.4.2: Replaced with combinator

Deprecated, see combinator

Re-exports

pub use binary::length_count;

pub use binary::length_data;

pub use binary::length_value;

Functions

count

Accumulate the output of a parser into a container, like Vec

fill

Repeats the embedded parser, filling the given slice with results.

fold_many0Deprecated

Deprecated, replaced by combinator::fold_repeat0

fold_many1Deprecated

Deprecated, replaced by combinator::fold_repeat1

fold_many_m_nDeprecated

Deprecated, replaced by combinator::fold_repeat

many0Deprecated

Deprecated, replaced by combinator::repeat0

many1Deprecated

Deprecated, replaced by combinator::repeat1

many_m_nDeprecated

Deprecated, replaced by combinator::repeat

many_till0Deprecated

Deprecated, replaced by combinator::repeat_till0

separated0

Accumulate the output of a parser, interleaed with sep

separated1

Accumulate the output of a parser, interleaed with sep

separated_foldl1

Alternates between two parsers, merging the results (left associative)

separated_foldr1

Alternates between two parsers, merging the results (right associative)