Function winnow::multi::many0

pub fn many0<I, O, C, E, F>(f: F) -> impl Parser<I, C, E>where
    I: Stream,
    C: Accumulate<O>,
    F: Parser<I, O, E>,
    E: ParseError<I>,

Accumulate the output of a parser into a container, like Vec

This stops on ErrMode::Backtrack. To instead chain an error up, see cut_err.

To recognize a series of tokens, Accumulate into a () and then Parser::recognize.

Warning: if the parser passed in accepts empty inputs (like alpha0 or digit0), many0 will return an error, to prevent going into an infinite loop

Example

use winnow::multi::many0;
use winnow::bytes::tag;

fn parser(s: &str) -> IResult<&str, Vec<&str>> {
  many0("abc").parse_next(s)
}

assert_eq!(parser("abcabc"), Ok(("", vec!["abc", "abc"])));
assert_eq!(parser("abc123"), Ok(("123", vec!["abc"])));
assert_eq!(parser("123123"), Ok(("123123", vec![])));
assert_eq!(parser(""), Ok(("", vec![])));