Module winnow::sequence

Combinators applying parsers in sequence

Functions

delimited

Apply three parsers, only returning the output of the second.

preceded

Apply two parsers, only returning the output from the second.

separated_pair

Apply three parsers, only returning the values of the first and third.

terminated

Apply two parsers, only returning the output of the first.