Trait winnow::sequence::Tuple

pub trait Tuple<I, O, E> {
    // Required method
    fn parse(&mut self, input: I) -> IResult<I, O, E>;
}

👎Deprecated since 0.1.0: Replaced with Parser

Helper trait for the tuple combinator.

This trait is implemented for tuples of parsers of up to 21 elements.

Required Methods

fn parse(&mut self, input: I) -> IResult<I, O, E>

👎Deprecated since 0.1.0: Replaced with Parser

Parses the input and returns a tuple of results of each parser.