Struct winnow::stream::Located

pub struct Located<I> { /* private fields */ }

Allow collecting the span of a parsed token

Spans are tracked as a Range<usize> of byte offsets.

Converting byte offsets to line or column numbers is left up to the user, as computing column numbers requires domain knowledge (are columns byte-based, codepoint-based, or grapheme-based?) and O(n) iteration over the input to determine codepoint and line boundaries.

The line-span crate can help with converting byte offsets to line numbers.

See Parser::span and Parser::with_span for more details

Implementations

impl<I> Located<I>
where I: Clone + Offset,

pub fn new(input: I) -> Self

Wrap another Stream with span tracking