winnow::stream

Struct LocatingSlice

pub struct LocatingSlice<I> { /* private fields */ }

Allow collecting the span of a parsed token within a slice

Converting byte offsets to line or column numbers is left up to the user, as computing column numbers requires domain knowledge (are columns byte-based, codepoint-based, or grapheme-based?) and O(n) iteration over the input to determine codepoint and line boundaries.

The line-span crate can help with converting byte offsets to line numbers.

See Parser::span and Parser::with_span for more details

Implementations

impl<I> LocatingSlice<I>
where I: Clone + Offset,

pub fn new(input: I) -> Self

Wrap another Stream with span tracking

impl<I> LocatingSlice<I>
where I: Clone + Stream + Offset,

pub fn reset_to_start(&mut self)

Reset the stream to the start

This is useful for formats that encode a graph with addresses relative to the start of the input.