Trait winnow::stream::HexDisplay

pub trait HexDisplay {
    // Required methods
    fn to_hex(&self, chunk_size: usize) -> String;
    fn to_hex_from(&self, chunk_size: usize, from: usize) -> String;
}

👎Deprecated since 0.3.0: Replaced with Debug (see Bytes and BStr

Deprecated, replaced with Debug, particularly Bytes and BStr

Required Methods

fn to_hex(&self, chunk_size: usize) -> String

👎Deprecated since 0.3.0: Replaced with Debug (see Bytes and BStr

Converts the value of self to a hex dump, returning the owned String.

fn to_hex_from(&self, chunk_size: usize, from: usize) -> String

👎Deprecated since 0.3.0: Replaced with Debug (see Bytes and BStr

Converts the value of self to a hex dump beginning at from address, returning the owned String.