Trait winnow::stream::Offset

pub trait Offset {
    // Required method
    fn offset_to(&self, second: &Self) -> usize;
}

Useful functions to calculate the offset between slices and show a hexdump of a slice

Required Methods

fn offset_to(&self, second: &Self) -> usize

Offset between the first byte of self and the first byte of the argument