Module winnow::token

Parsers extracting tokens from the stream

Functions

any

Matches one token

none_of

Recognize a token that does not match the pattern

one_of

Recognize a token that matches the pattern

tag

Recognizes a literal

tag_no_case

Recognizes a case insensitive literal.

take

Recognize an input slice containing the first N input elements (I[..N]).

take_till0

Recognize the longest input slice (if any) till a pattern is met.

take_till1

Recognize the longest (at least 1) input slice till a pattern is met.

take_until0

Recognize the input slice up to the first occurrence of the literal.

take_until1

Recognize the non empty input slice up to the first occurrence of the literal.

take_while

Recognize the longest (m <= len <= n) input slice that matches the pattern

take_while0

Recognize the longest input slice (if any) that matches the pattern

take_while1

Recognize the longest (at least 1) input slice that matches the pattern