Function winnow::token::take_till0_complete

fn take_till0_complete<P, I: Stream, E: ParserError<I>>(
    input: &mut I,
    predicate: P
) -> PResult<<I as Stream>::Slice, E>
where
    P: Fn(I::Token) -> bool,

Looks for the first element of the input type for which the condition returns true, and returns the input up to this position.

Complete version: If no element is found matching the condition, this will return the whole input