winnow

Module ascii

Character specific parsers and combinators

Functions recognizing specific characters

Structs

Caseless

Mark a value as case-insensitive for ASCII characters

Traits

HexUint

Metadata for parsing hex numbers, see hex_uint

Int

Metadata for parsing signed integers, see dec_int

Uint

Metadata for parsing unsigned integers, see dec_uint

Functions

alpha0

Recognizes zero or more lowercase and uppercase ASCII alphabetic characters: 'a'..='z', 'A'..='Z'

alpha1

Recognizes one or more lowercase and uppercase ASCII alphabetic characters: 'a'..='z', 'A'..='Z'

alphanumeric0

Recognizes zero or more ASCII numerical and alphabetic characters: 'a'..='z', 'A'..='Z', '0'..='9'

alphanumeric1

Recognizes one or more ASCII numerical and alphabetic characters: 'a'..='z', 'A'..='Z', '0'..='9'

crlf

Recognizes the string "\r\n".

dec_int

Decode a decimal signed integer (e.g. i32)

dec_uint

Decode a decimal unsigned integer (e.g. u32)

digit0

Recognizes zero or more ASCII numerical characters: '0'..='9'

digit1

Recognizes one or more ASCII numerical characters: '0'..='9'

escaped

Parse escaped characters, unescaping them

escaped_transformDeprecated

Deprecated, replaed with escaped

float

Recognizes floating point number in text format and returns a f32 or f64.

hex_digit0

Recognizes zero or more ASCII hexadecimal numerical characters: '0'..='9', 'A'..='F', 'a'..='f'

hex_digit1

Recognizes one or more ASCII hexadecimal numerical characters: '0'..='9', 'A'..='F', 'a'..='f'

hex_uint

Decode a variable-width hexadecimal integer (e.g. u32)

line_ending

Recognizes an end of line (both "\n" and "\r\n").

multispace0

Recognizes zero or more spaces, tabs, carriage returns and line feeds.

multispace1

Recognizes one or more spaces, tabs, carriage returns and line feeds.

newline

Matches a newline character '\n'.

oct_digit0

Recognizes zero or more octal characters: '0'..='7'

oct_digit1

Recognizes one or more octal characters: '0'..='7'

space0

Recognizes zero or more spaces and tabs.

space1

Recognizes one or more spaces and tabs.

tab

Matches a tab character '\t'.

take_escaped

Recognize the input slice with escaped characters.

till_line_ending

Recognizes a string of any char except "\r\n" or "\n".