Module winnow::ascii

Character specific parsers and combinators

Functions recognizing specific characters

Traits

HexUint

Metadata for parsing hex numbers, see hex_uint

Int

Metadata for parsing signed integers, see dec_int

Uint

Metadata for parsing unsigned integers, see dec_uint

Functions

alpha0

Recognizes zero or more lowercase and uppercase ASCII alphabetic characters: a-z, A-Z

alpha1

Recognizes one or more lowercase and uppercase ASCII alphabetic characters: a-z, A-Z

alphanumeric0

Recognizes zero or more ASCII numerical and alphabetic characters: 0-9, a-z, A-Z

alphanumeric1

Recognizes one or more ASCII numerical and alphabetic characters: 0-9, a-z, A-Z

crlf

Recognizes the string “\r\n”.

dec_int

Decode a decimal signed integer

dec_uint

Decode a decimal unsigned integer

digit0

Recognizes zero or more ASCII numerical characters: 0-9

digit1

Recognizes one or more ASCII numerical characters: 0-9

escaped

Matches a byte string with escaped characters.

escaped_transform

Matches a byte string with escaped characters.

float

Recognizes floating point number in text format and returns a f32 or f64.

hex_digit0

Recognizes zero or more ASCII hexadecimal numerical characters: 0-9, A-F, a-f

hex_digit1

Recognizes one or more ASCII hexadecimal numerical characters: 0-9, A-F, a-f

hex_uint

Decode a variable-width hexadecimal integer.

line_ending

Recognizes an end of line (both ‘\n’ and ‘\r\n’).

multispace0

Recognizes zero or more spaces, tabs, carriage returns and line feeds.

multispace1

Recognizes one or more spaces, tabs, carriage returns and line feeds.

newline

Matches a newline character ‘\n’.

not_line_ending

Recognizes a string of any char except ‘\r\n’ or ‘\n’.

oct_digit0

Recognizes zero or more octal characters: 0-7

oct_digit1

Recognizes one or more octal characters: 0-7

space0

Recognizes zero or more spaces and tabs.

space1

Recognizes one or more spaces and tabs.

tab

Matches a tab character ‘\t’.