Function winnow::ascii::space1

pub fn space1<I, E: ParserError<I>>(
    input: &mut I
) -> PResult<<I as Stream>::Slice, E>
where
    I: StreamIsPartial + Stream,
    <I as Stream>::Token: AsChar + Clone,

Recognizes zero or more spaces and tabs.

Complete version: Will return the whole input if no terminating token is found (a non space character).

Partial version: Will return Err(winnow::error::ErrMode::Incomplete(_)) if there’s not enough input data, or if no terminating token is found (a non space character).

Example

fn parser<'s>(input: &mut &'s str) -> PResult<&'s str, InputError<&'s str>> {
    space1.parse_next(input)
}

assert_eq!(parser.parse_peek(" \t21c"), Ok(("21c", " \t")));
assert_eq!(parser.parse_peek("H2"), Err(ErrMode::Backtrack(InputError::new("H2", ErrorKind::Slice))));
assert_eq!(parser.parse_peek(""), Err(ErrMode::Backtrack(InputError::new("", ErrorKind::Slice))));

assert_eq!(space1::<_, InputError<_>>.parse_peek(Partial::new(" \t21c")), Ok((Partial::new("21c"), " \t")));
assert_eq!(space1::<_, InputError<_>>.parse_peek(Partial::new("H2")), Err(ErrMode::Backtrack(InputError::new(Partial::new("H2"), ErrorKind::Slice))));
assert_eq!(space1::<_, InputError<_>>.parse_peek(Partial::new("")), Err(ErrMode::Incomplete(Needed::new(1))));