Module winnow::binary::bits

Bit level parsers

Functions

bits

Converts a byte-level input to a bit-level input

bool

Parses one specific bit as a bool.

bytes

Convert a bits stream back into a byte stream

pattern

Parse taking count bits and comparing them to pattern

tagDeprecated

Deprecated, replaced with pattern

take

Parse taking count bits