Function winnow::binary::length_repeat

pub fn length_repeat<I, O, C, N, E, F, G>(f: F, g: G) -> impl Parser<I, C, E>
where
    I: Stream,
    N: ToUsize,
    C: Accumulate<O>,
    F: Parser<I, N, E>,
    G: Parser<I, O, E>,
    E: ParserError<I>,

Accumulate a length-prefixed sequence of values (TLV)

If the length represents token counts, see instead length_take

Arguments

f The parser to apply to obtain the count.

g The parser to apply repeatedly.

Example

use winnow::Bytes;
use winnow::binary::u8;
use winnow::binary::length_count;
use winnow::token::tag;

type Stream<'i> = &'i Bytes;

fn stream(b: &[u8]) -> Stream<'_> {
    Bytes::new(b)
}

fn parser(s: Stream<'_>) -> IResult<Stream<'_>, Vec<&[u8]>> {
  length_count(u8.map(|i| {
     println!("got number: {}", i);
     i
  }), "abc").parse_peek(s)
}

assert_eq!(parser(stream(b"\x02abcabcabc")), Ok((stream(b"abc"), vec![&b"abc"[..], &b"abc"[..]])));
assert_eq!(parser(stream(b"\x03123123123")), Err(ErrMode::Backtrack(InputError::new(stream(b"123123123"), ErrorKind::Tag))));