winnow::combinator

Function separated

pub fn separated<Input, Output, Accumulator, Sep, Error, ParseNext, SepParser>(
    occurrences: impl Into<Range>,
    parser: ParseNext,
    separator: SepParser,
) -> impl Parser<Input, Accumulator, Error>
where
    Input: Stream,
    Accumulator: Accumulate<Output>,
    ParseNext: Parser<Input, Output, Error>,
    SepParser: Parser<Input, Sep, Error>,
    Error: ParserError<Input>,

Accumulate the output of a parser, interleaved with sep

This stops when either parser returns ErrMode::Backtrack. To instead chain an error up, see cut_err.

To take a series of tokens, Accumulate into a () (e.g. with .map(|()| ())) and then Parser::take.

Warning: If the separator parser accepts empty inputs (like alpha0 or digit0), separated will return an error, to prevent going into an infinite loop.

Example

Zero or more repetitions:

use winnow::combinator::separated;

fn parser<'i>(s: &mut &'i str) -> ModalResult<Vec<&'i str>> {
  separated(0.., "abc", "|").parse_next(s)
}

assert_eq!(parser.parse_peek("abc|abc|abc"), Ok(("", vec!["abc", "abc", "abc"])));
assert_eq!(parser.parse_peek("abc123abc"), Ok(("123abc", vec!["abc"])));
assert_eq!(parser.parse_peek("abc|def"), Ok(("|def", vec!["abc"])));
assert_eq!(parser.parse_peek(""), Ok(("", vec![])));
assert_eq!(parser.parse_peek("def|abc"), Ok(("def|abc", vec![])));

One or more repetitions:

use winnow::combinator::separated;

fn parser<'i>(s: &mut &'i str) -> ModalResult<Vec<&'i str>> {
  separated(1.., "abc", "|").parse_next(s)
}

assert_eq!(parser.parse_peek("abc|abc|abc"), Ok(("", vec!["abc", "abc", "abc"])));
assert_eq!(parser.parse_peek("abc123abc"), Ok(("123abc", vec!["abc"])));
assert_eq!(parser.parse_peek("abc|def"), Ok(("|def", vec!["abc"])));
assert!(parser.parse_peek("").is_err());
assert!(parser.parse_peek("def|abc").is_err());

Fixed number of repetitions:

use winnow::combinator::separated;

fn parser<'i>(s: &mut &'i str) -> ModalResult<Vec<&'i str>> {
  separated(2, "abc", "|").parse_next(s)
}

assert_eq!(parser.parse_peek("abc|abc|abc"), Ok(("|abc", vec!["abc", "abc"])));
assert!(parser.parse_peek("abc123abc").is_err());
assert!(parser.parse_peek("abc|def").is_err());
assert!(parser.parse_peek("").is_err());
assert!(parser.parse_peek("def|abc").is_err());

Arbitrary repetitions:

use winnow::combinator::separated;

fn parser<'i>(s: &mut &'i str) -> ModalResult<Vec<&'i str>> {
  separated(0..=2, "abc", "|").parse_next(s)
}

assert_eq!(parser.parse_peek("abc|abc|abc"), Ok(("|abc", vec!["abc", "abc"])));
assert_eq!(parser.parse_peek("abc123abc"), Ok(("123abc", vec!["abc"])));
assert_eq!(parser.parse_peek("abc|def"), Ok(("|def", vec!["abc"])));
assert_eq!(parser.parse_peek(""), Ok(("", vec![])));
assert_eq!(parser.parse_peek("def|abc"), Ok(("def|abc", vec![])));