Module winnow::combinator::multi

Combinators applying their child parser multiple times

Structs

Repeat

Implementation of repeat

Functions

fill

Repeats the embedded parser, filling the given slice with results.

fold_repeatDeprecated

Deprecated, replaced with Repeat::fold

fold_repeat0_ 🔒

fold_repeat1_ 🔒

fold_repeat_m_n_ 🔒

repeat

Accumulate the output of a parser into a container, like Vec

repeat0_ 🔒

repeat1_ 🔒

repeat_m_n_ 🔒

repeat_n_ 🔒

repeat_till

Accumulate the output of parser f into a container, like Vec, until the parser g produces a result.

repeat_till0Deprecated

Deprecated, replaced with repeat_till

repeat_till0_ 🔒

repeat_till_m_n_ 🔒

separated

Accumulate the output of a parser, interleaved with sep

separated0Deprecated

Accumulate the output of a parser, interleaved with sep

separated0_ 🔒

separated1Deprecated

Accumulate the output of a parser, interleaved with sep

separated1_ 🔒

separated_foldl1

Alternates between two parsers, merging the results (left associative)

separated_foldr1

Alternates between two parsers, merging the results (right associative)

separated_m_n_ 🔒

separated_n_ 🔒