Module winnow::combinator::multi

Combinators applying their child parser multiple times

Structs

Repeat

Implementation of repeat

Functions

fill

Repeats the embedded parser, filling the given slice with results.

fold_repeatDeprecated

Deprecated, replaced with Repeat::fold

fold_repeat0_ ğŸ”’

fold_repeat1_ ğŸ”’

fold_repeat_m_n_ ğŸ”’

repeat

Accumulate the output of a parser into a container, like Vec

repeat0_ ğŸ”’

repeat1_ ğŸ”’

repeat_m_n_ ğŸ”’

repeat_n_ ğŸ”’

repeat_till

Accumulate the output of parser f into a container, like Vec, until the parser g produces a result.

repeat_till0Deprecated

Deprecated, replaced with repeat_till

repeat_till0_ ğŸ”’

repeat_till_m_n_ ğŸ”’

separated

Accumulate the output of a parser, interleaved with sep

separated0Deprecated

Accumulate the output of a parser, interleaved with sep

separated0_ ğŸ”’

separated1Deprecated

Accumulate the output of a parser, interleaved with sep

separated1_ ğŸ”’

separated_foldl1

Alternates between two parsers, merging the results (left associative)

separated_foldr1

Alternates between two parsers, merging the results (right associative)

separated_m_n_ ğŸ”’

separated_n_ ğŸ”’