Struct winnow::error::InputError

pub struct InputError<I: Clone> {
    pub input: I,
    pub kind: ErrorKind,
}

Capture input on error

This is useful for testing of generic parsers to ensure the error happens at the right location.

Note: context and inner errors (like from Parser::try_map) will be dropped.

Fields
input: I

The input stream, pointing to the location where the error occurred

kind: ErrorKind

A rudimentary error kind

Implementations

impl<I: Clone> InputError<I>

pub fn new(input: I, kind: ErrorKind) -> Self

Creates a new basic error

pub fn map_input<I2: Clone, O: Fn(I) -> I2>(self, op: O) -> InputError<I2>

Translate the input type

impl<'i, I: ToOwned> InputError<&'i I>
where <I as ToOwned>::Owned: Clone,

pub fn into_owned(self) -> InputError<<I as ToOwned>::Owned>

Obtaining ownership