winnow::error

Trait AddContext

pub trait AddContext<I: Stream, C = &'static str>: Sized {
    // Provided method
    fn add_context(
        self,
        _input: &I,
        _token_start: &<I as Stream>::Checkpoint,
        _context: C,
    ) -> Self { ... }
}

Used by Parser::context to add custom data to error while backtracking

May be implemented multiple times for different kinds of context.

Provided Methods

fn add_context( self, _input: &I, _token_start: &<I as Stream>::Checkpoint, _context: C, ) -> Self

Append to an existing error custom data

This is used mainly by Parser::context, to add user friendly information to errors when backtracking through a parse tree

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.