Trait winnow::error::AddContext

pub trait AddContext<I, C = &'static str>: Sized {
    // Provided method
    fn add_context(self, _input: &I, _ctx: C) -> Self { ... }
}

Used by Parser::context to add custom data to error while backtracking

May be implemented multiple times for different kinds of context.

Provided Methods

fn add_context(self, _input: &I, _ctx: C) -> Self

Append to an existing error custom data

This is used mainly by Parser::context, to add user friendly information to errors when backtracking through a parse tree

Object Safety

This trait is not object safe.