Trait winnow::error::ParseError

pub trait ParseError<I>: Sized {
    // Required methods
    fn from_error_kind(input: I, kind: ErrorKind) -> Self;
    fn append(self, input: I, kind: ErrorKind) -> Self;

    // Provided methods
    fn assert(input: I, _message: &'static str) -> Self
       where I: Debug { ... }
    fn from_char(input: I, _: char) -> Self { ... }
    fn or(self, other: Self) -> Self { ... }
}

The basic Parser trait for errors

It provides methods to create an error from some combinators, and combine existing errors in combinators like alt.

Required Methods

fn from_error_kind(input: I, kind: ErrorKind) -> Self

Creates an error from the input position and an ErrorKind

fn append(self, input: I, kind: ErrorKind) -> Self

Like ParseError::from_error_kind but merges it with the existing error.

This is useful when backtracking through a parse tree, accumulating error context on the way.

Provided Methods

fn assert(input: I, _message: &'static str) -> Selfwhere I: Debug,

Process a parser assertion

fn from_char(input: I, _: char) -> Self

👎Deprecated since 0.2.0: Replaced with ContextError

Creates an error from an input position and an expected character

fn or(self, other: Self) -> Self

Combines errors from two different parse branches.

For example, this would be used by alt to report the error from each case.