Type Definition winnow::error::IResult

pub type IResult<I, O, E = Error<I>> = Result<(I, O), ErrMode<E>>;

Holds the result of Parser

Ok((I, O)) is the remaining input and the parsed value

Err(ErrMode<E>) is the error along with how to respond to it

By default, the error type (E) is Error

At the top-level of your parser, you can use the FinishIResult::finish method to convert it to a more common result type