Type Alias winnow::error::PResult

pub type PResult<O, E = ContextError> = Result<O, ErrMode<E>>;

For use with Parser::parse_next

Ok(O) is the parsed value

Err(ErrMode<E>) is the error along with how to respond to it

By default, the error type (E) is ContextError.

When integrating into the result of the application, see

Parser::parse

ErrMode::into_inner

Aliased Type

enum PResult<O, E = ContextError> {
    Ok(O),
    Err(ErrMode<E>),
}

Variants

Ok(O)

Contains the success value

Err(ErrMode<E>)

Contains the error value