Module winnow::multi

Combinators applying their child parser multiple times

Functions

count

Accumulate the output of a parser into a container, like Vec

fill

Runs the embedded parser repeatedly, filling the given slice with results.

fold_many0

Repeats the embedded parser, calling g to gather the results.

fold_many1

Repeats the embedded parser, calling g to gather the results.

fold_many_m_n

Repeats the embedded parser m..=n times, calling g to gather the results

length_count

Gets a number from the first parser, then applies the second parser that many times.

length_data

Gets a number from the parser and returns a subslice of the input of that size.

length_value

Gets a number from the first parser, takes a subslice of the input of that size, then applies the second parser on that subslice. If the second parser returns Incomplete, length_value will return an error.

many0

Accumulate the output of a parser into a container, like Vec

many1

Accumulate the output of a parser into a container, like Vec

many_m_n

Repeats the embedded parser m..=n times

many_till0

Applies the parser f until the parser g produces a result.

separated0

Alternates between two parsers to produce a list of elements.

separated1

Alternates between two parsers to produce a list of elements until ErrMode::Backtrack.

separated_foldl1

Alternates between two parsers, merging the results (left associative)

separated_foldr1

Alternates between two parsers, merging the results (right associative)