winnow::stream

Struct Partial

pub struct Partial<I> { /* private fields */ }

Mark the input as a partial buffer for streaming input.

Complete input means that we already have all of the data. This will be the common case with small files that can be read entirely to memory.

In contrast, streaming input assumes that we might not have all of the data. This can happen with some network protocol or large file parsers, where the input buffer can be full and need to be resized or refilled.

ErrMode::Incomplete will report how much more data is needed.

Parser::complete_err transform ErrMode::Incomplete to ErrMode::Backtrack

See also StreamIsPartial to tell whether the input supports complete or partial parsing.

See also [Special Topics: Parsing Partial Input][crate::_topic::partial].

Example

Here is how it works in practice:

fn take_partial<'s>(i: &mut Partial<&'s [u8]>) -> ModalResult<&'s [u8], ContextError> {
  token::take(4u8).parse_next(i)
}

fn take_complete<'s>(i: &mut &'s [u8]) -> ModalResult<&'s [u8], ContextError> {
  token::take(4u8).parse_next(i)
}

// both parsers will take 4 bytes as expected
assert_eq!(take_partial.parse_peek(Partial::new(&b"abcde"[..])), Ok((Partial::new(&b"e"[..]), &b"abcd"[..])));
assert_eq!(take_complete.parse_peek(&b"abcde"[..]), Ok((&b"e"[..], &b"abcd"[..])));

// if the input is smaller than 4 bytes, the partial parser
// will return `Incomplete` to indicate that we need more data
assert_eq!(take_partial.parse_peek(Partial::new(&b"abc"[..])), Err(ErrMode::Incomplete(Needed::new(1))));

// but the complete parser will return an error
assert!(take_complete.parse_peek(&b"abc"[..]).is_err());

// the alpha0 function takes 0 or more alphabetic characters
fn alpha0_partial<'s>(i: &mut Partial<&'s str>) -> ModalResult<&'s str, ContextError> {
  ascii::alpha0.parse_next(i)
}

fn alpha0_complete<'s>(i: &mut &'s str) -> ModalResult<&'s str, ContextError> {
  ascii::alpha0.parse_next(i)
}

// if there's a clear limit to the taken characters, both parsers work the same way
assert_eq!(alpha0_partial.parse_peek(Partial::new("abcd;")), Ok((Partial::new(";"), "abcd")));
assert_eq!(alpha0_complete.parse_peek("abcd;"), Ok((";", "abcd")));

// but when there's no limit, the partial version returns `Incomplete`, because it cannot
// know if more input data should be taken. The whole input could be "abcd;", or
// "abcde;"
assert_eq!(alpha0_partial.parse_peek(Partial::new("abcd")), Err(ErrMode::Incomplete(Needed::new(1))));

// while the complete version knows that all of the data is there
assert_eq!(alpha0_complete.parse_peek("abcd"), Ok(("", "abcd")));

Implementations

impl<I> Partial<I>
where I: StreamIsPartial,

pub fn new(input: I) -> Self

Create a partial input

pub fn into_inner(self) -> I

Extract the original Stream