winnow::stream

Trait Location

pub trait Location {
    // Required method
    fn location(&self) -> usize;
}

Number of indices input has advanced since start of parsing

See LocatingSlice for adding location tracking to your Stream

Required Methods

fn location(&self) -> usize

Number of indices input has advanced since start of parsing