winnow::stream

Trait Offset

pub trait Offset<Start = Self> {
    // Required method
    fn offset_from(&self, start: &Start) -> usize;
}

Useful functions to calculate the offset between slices and show a hexdump of a slice

Required Methods

fn offset_from(&self, start: &Start) -> usize

Offset between the first byte of start and the first byte of selfa

Note: This is an offset, not an index, and may point to the end of input (start.len()) when self is exhausted.