Module winnow::token

Parsers extracting tokens from the stream

Functions

any

Matches one token

literal

Recognizes a literal

none_of

Recognize a token that does not match the pattern

one_of

Recognize a token that matches the pattern

tagDeprecated

Deprecated, replaced with literal

tag_no_caseDeprecated

Recognizes a case insensitive literal.

take

Recognize an input slice containing the first N input elements (I[..N]).

take_till

Recognize the longest input slice (if any) till a pattern is met.

take_till0Deprecated

Recognize the longest input slice (if any) till a pattern is met.

take_till1Deprecated

Recognize the longest (at least 1) input slice till a pattern is met.

take_until

Recognize the input slice up to the first occurrence of the literal.

take_until0Deprecated

Deprecated, see take_until

take_until1Deprecated

Deprecated, see take_until

take_while

Recognize the longest (m <= len <= n) input slice that matches the pattern